pyproject.toml
[project]
name = "event-manager-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "authlib>=1.3.2",
    "fastapi>=0.117.1",
    "flask>=3.0.3",
    "google-auth>=2.40.3",
    "httpx>=0.28.1",
    "itsdangerous>=2.2.0",
    "mysql-connector-python>=9.0.0",
    "psycopg2-binary>=2.9.10",
    "pyjwt>=2.9.0",
    "python-dotenv>=1.0.1",
    "python-jose>=3.4.0",
    "requests>=2.32.4",
    "sqlalchemy>=2.0.44",
    "starlette>=0.44.0",
    "uvicorn>=0.33.0",
    "werkzeug>=3.0.6",
]
